+++ react/components/ProductPrice/propTypes.js
@@ -2,10 +2,18 @@ import PropTypes from 'prop-types'
 import { intlShape } from 'react-intl'
 
 export default {
+  /** Product list of selling prices */
+  sellingPriceList: PropTypes.arrayOf(PropTypes.number),
+  /** Product list of list prices */
+  listPriceList: PropTypes.arrayOf(PropTypes.number),
   /** Product selling price */
   sellingPrice: PropTypes.number,
   /** Product list price */
   listPrice: PropTypes.number,
+  /** Set visibility of selling prices range */
+  showSellingPriceRange: PropTypes.bool.isRequired,
+  /** Set visibility of list prices range */
+  showListPriceRange: PropTypes.bool.isRequired,
   /** Set visibility of list price */
   showListPrice: PropTypes.bool.isRequired,
   /** Set visibility of labels */
@@ -43,6 +51,10 @@ export default {
   listPriceLabelClass: PropTypes.string,
   /** Classes to be applied to price value */
   listPriceClass: PropTypes.string,
+  /** Classes to be applied to range selling prices value */
+  sellingPriceRangeClass: PropTypes.string,
+  /** Classes to be applied to range list prices value */
+  listPriceRangeClass: PropTypes.string,
   /** Classes to be applied to selling price container */
   sellingPriceContainerClass: PropTypes.string,
   /** Classes to be applied to selling price label */